pyproject.toml
[project]
name = "static-workflow-rag-cli"
version = "0.1.0"
description = "Static workflow RAG CLI for local document QA"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "chromadb>=0.5.0",
  "rank-bm25>=0.2.2",
  "python-docx>=1.1.0",
  "requests>=2.32.0",
]

[project.scripts]
rag-cli = "app.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
